% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_flight_path.R
\name{make_flight_path}
\alias{make_flight_path}
\title{Make flight path plot}
\usage{
make_flight_path(x, y, country = "Nepal")
}
\arguments{
\item{x}{Longitude coordinates of start and end point}

\item{y}{Latitude coordinates of start and end point}
}
\value{
plot
}
\description{
Create a flight path plot
}
