% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_flight_path_elevation_plot.R
\name{make_flight_path_elevation_plot}
\alias{make_flight_path_elevation_plot}
\title{Make flight path elevation plot}
\usage{
make_flight_path_elevation_plot(df)
}
\arguments{
\item{df}{a dataframe generated from make_flight_path}
}
\value{
plot
}
\description{
Create a flight path elevation plot
}
