% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plan.R
\name{read_plan}
\alias{read_plan}
\title{read_plan}
\usage{
read_plan(plan_file = "Pyuthan to Dharmawati.plan", return_name = TRUE)
}
\arguments{
\item{plan_file}{The path to the file}

\item{return_name}{Whether to parse the filename to include in the output}
}
\description{
Read a .plan file
}
