% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_db.R
\name{update_db}
\alias{update_db}
\title{Update DB}
\usage{
update_db(data, table_name = "transcriptions",
  connection_object = NULL)
}
\arguments{
\item{data}{The data to be uploaded}

\item{table_name}{Name of table}

\item{connection_object}{The connection object}
}
\value{
The transcriptions relation in the postgresql database will be updated
}
\description{
Update a table
}
\examples{
2+2
}
