% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_table.R
\name{write_table}
\alias{write_table}
\title{Write table}
\usage{
write_table(connection_object = NULL, table = "transcriptions", value)
}
\arguments{
\item{connection_object}{An open connection to adatabase (as created through \code{credentials_extract} and \code{credentials_connect} or \code{credentials_now}); if \code{NULL}, the function will try to create a \code{connection_object} by retrieving user information from the \code{credentials/credentials.yaml}
in or somewhere upwards of the working directory.}

\item{table}{The name of the table in the database to be written}

\item{value}{The data to be written}
}
\value{
A table will be written or overwritten
}
\description{
Write a table in the database
}
